+++ project2/geom2.h
@@ -18,11 +18,11 @@ using namespace std;
 // Handy geometry classes and functions
 
 // computes the distance between two two-dimensional points
-int distance(int xA, int yA, int xB, int yB);
+double distance(double xA, double yA, double xB, double yB);
 
 // computes the hypotenuse of a triangle formed by the x and y
 // values of a velocity
-int pythagorean(int x, int y);
+double pythagorean(double x, double y);
 
 // Conversion between degrees and radians
 float RadtoDeg(float rad);
@@ -32,16 +32,16 @@ float DegtoRad(float deg);
 class Velocity {
  public:
   Velocity();
-  Velocity(int x, int y);
+  Velocity(double x, double y);
   double getX() { return x; }
   double getY() { return y; }
   double getMagnitude() { return magnitude; }
   double getDirection() { return directionRad; }
   bool moving(double t);
-  void setXY(int x, int y);
+  void setXY(double x, double y);
   void slow(double t);
   void reverse();
-  void scalarSelfProduct(int scalar);
+  void scalarSelfProduct(double scalar);
   void print();
 
   // operators
@@ -57,14 +57,14 @@ class Velocity {
 };
 
 
-int DotProduct(Velocity vA, Velocity vB);
+double DotProduct(Velocity vA, Velocity vB);
 
 // Uses the dot product between two velocity vectors to compute
 // the angle between them. Returns an angle in radians.
 float AngleBetween(Velocity vA, Velocity vB);
 
 Velocity sumOfVelocities( Velocity vA, Velocity vB);
-Velocity scalarProduct(double scalar, Velocity v);
+Velocity scalarProduct( double scalar, Velocity v);
 
 
 #endif
